refactor code of solov2
