--- configs/solov2/solov2_light_r50_vd_fpn_dcn_512_3x.yml
@@ -1,68 +1,0 @@
-architecture: SOLOv2
-use_gpu: true
-max_iters: 270000
-snapshot_iter: 30000
-log_smooth_window: 20
-save_dir: output
-pretrain_weights: https://paddle-imagenet-models-name.bj.bcebos.com/ResNet50_vd_ssld_pretrained.tar
-metric: COCO
-weights: output/solov2_light_r50_vd_fpn_dcn_512_multiscale_3x/model_final
-num_classes: 81
-use_ema: true
-ema_decay: 0.9998
-
-SOLOv2:
-  backbone: ResNet
-  fpn: FPN
-  bbox_head: SOLOv2Head
-  mask_head: SOLOv2MaskHead
-
-ResNet:
-  depth: 50
-  feature_maps: [2, 3, 4, 5]
-  freeze_at: 2
-  norm_type: bn
-  dcn_v2_stages: [3, 4, 5]
-  variant: d
-  lr_mult_list: [0.05, 0.05, 0.1, 0.15]
-
-FPN:
-  max_level: 6
-  min_level: 2
-  num_chan: 256
-  spatial_scale: [0.03125, 0.0625, 0.125, 0.25]
-  reverse_out: True
-
-SOLOv2Head:
-  seg_feat_channels: 256
-  stacked_convs: 3
-  num_grids: [40, 36, 24, 16, 12]
-  kernel_out_channels: 128
-  dcn_v2_stages: [2,]
-  drop_block: True
-
-SOLOv2MaskHead:
-  out_channels: 128
-  start_level: 0
-  end_level: 3
-  num_classes: 128
-
-LearningRate:
-  base_lr: 0.01
-  schedulers:
-  - !PiecewiseDecay
-    gamma: 0.1
-    milestones: [180000, 240000]
-  - !LinearWarmup
-    start_factor: 0.
-    steps: 1000
-
-OptimizerBuilder:
-  optimizer:
-    momentum: 0.9
-    type: Momentum
-  regularizer:
-    factor: 0.0001
-    type: L2
-
-_READER_: 'solov2_light_reader.yml'
