+++ configs/solov2/solov2_r50_fpn_1x.yml
@@ -33,12 +33,23 @@ SOLOv2Head:
   stacked_convs: 4
   num_grids: [40, 36, 24, 16, 12]
   kernel_out_channels: 256
+  solov2_loss: SOLOv2Loss
+  mask_nms: MaskMatrixNMS
 
 SOLOv2MaskHead:
-  out_channels: 128
+  in_channels: 128
+  out_channels: 256
   start_level: 0
   end_level: 3
-  num_classes: 256
+
+SOLOv2Loss:
+  ins_loss_weight: 3.0
+  focal_loss_gamma: 2.0
+  focal_loss_alpha: 0.25
+
+MaskMatrixNMS:
+  pre_nms_top_n: 500
+  post_nms_top_n: 100
 
 LearningRate:
   base_lr: 0.01
