+++ ppdet/modeling/losses/__init__.py
@@ -24,6 +24,7 @@ from . import fcos_loss
 from . import diou_loss_yolo
 from . import iou_aware_loss
 from . import ssd_with_lmk_loss
+from . import solov2_loss
 
 from .iou_aware_loss import *
 from .yolo_loss import *
@@ -35,3 +36,4 @@ from .balanced_l1_loss import *
 from .fcos_loss import *
 from .diou_loss_yolo import *
 from .ssd_with_lmk_loss import *
+from .solov2_loss import *
